Rename jsWasmMain to webMain
