+++ compose/animation/animation-core/build.gradle
@@ -122,10 +122,10 @@ if (AndroidXComposePlugin.isMultiplatformEnabled(project)) {
             }
             jsNativeMain.dependsOn(commonMain)
             nativeMain.dependsOn(jsNativeMain)
-            jsWasmMain.dependsOn(jsNativeMain)
+            webMain.dependsOn(jsNativeMain)
 
             jsMain {
-                dependsOn(jsWasmMain)
+                dependsOn(webMain)
             }
 
             jbMain {
@@ -138,7 +138,7 @@ if (AndroidXComposePlugin.isMultiplatformEnabled(project)) {
             jsNativeMain.dependsOn(jbMain)
 
             wasmJsMain {
-                dependsOn(jsWasmMain)
+                dependsOn(webMain)
             }
 
             // TODO(b/214407011): These dependencies leak into instrumented tests as well. If you
