+++ compose/foundation/foundation-layout/build.gradle
@@ -109,7 +109,7 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
             jsNativeMain.dependsOn(commonMain)
             nativeMain.dependsOn(jsNativeMain)
 
-            jsWasmMain {
+            webMain {
                 dependsOn(jsNativeMain)
                 dependencies {
                     implementation(libs.kotlinStdlib)
@@ -121,14 +121,14 @@ if(AndroidXComposePlugin.isMultiplatformEnabled(project)) {
             notMobileMain.dependsOn(skikoMain)
             desktopMain.dependsOn(notMobileMain)
             macosMain.dependsOn(notMobileMain)
-            jsWasmMain.dependsOn(notMobileMain)
+            webMain.dependsOn(notMobileMain)
 
             jsMain {
-                dependsOn(jsWasmMain)
+                dependsOn(webMain)
             }
 
             wasmJsMain {
-                dependsOn(jsWasmMain)
+                dependsOn(webMain)
             }
 
             // TODO(b/214407011): These dependencies leak into instrumented tests as well. If you
