+++ compose/foundation/foundation/src/webMain/kotlin/androidx/compose/foundation/text/NativeCursorHandle.web.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2023 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package androidx.compose.foundation.text
+
+import androidx.compose.foundation.text.selection.OffsetProvider
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.DpSize
+
+@Composable
+@Suppress("UNUSED_PARAMETER")
+internal actual fun CursorHandle(
+    offsetProvider: OffsetProvider,
+    modifier: Modifier,
+    minTouchTargetSize: DpSize
+) {
+    /* Not implemented. */
+}
+
