+++ compose/foundation/foundation/src/webMain/kotlin/androidx/compose/foundation/text/TouchMode.web.kt
