+++ compose/foundation/foundation/src/webMain/kotlin/androidx/compose/foundation/text/contextmenu/internal/ProvideDefaultPlatformTextContextMenuProviders.web.kt
