+++ compose/foundation/foundation/src/webMain/kotlin/androidx/compose/foundation/text/selection/SelectionManager.web.kt
