+++ compose/ui/ui/src/webMain/kotlin/androidx/compose/ui/platform/GlobalSnapshotManager.web.kt
